project/pulsar-consumer/requirements.txt
pulsar-client
maven
PyGithub
requests

